requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
android==1.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
astroid==3.3.9
attrs==25.3.0
audiostream==0.2
blinker==1.9.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cmake @ file:///home/n0n3m4/Desktop/Pydroid/zBuildPrepared/cmake-3.31.5-py3-none-any.whl#sha256=0c372e231235ded65541b768b4450199bf9238b4e3dcdb8885f17c6d1f9007b6
dill==0.4.0
distro==1.9.0
Django==5.2.4
django-allauth==65.10.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
docutils==0.21.2
drf-yasg==1.21.10
filetype==1.2.0
Flask==3.1.1
frozenlist==1.7.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
isort==6.0.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
Kivy==2.3.1
lazy-object-proxy==0.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
multidict==6.6.3
ninja @ file:///home/n0n3m4/Desktop/Pydroid/zBuildPrepared/ninja-1.12.1-py3-none-any.whl#sha256=369aa5ff36a1bbe7299b02cd6d67f2589e7fd3e3871303c13bafa62a87b5b4fc
openai==1.97.0
packaging==25.0
parso==0.8.4
pkgconfig==1.5.5
platformdirs==4.3.7
pluggy==1.6.0
propcache==0.3.2
pybind11==2.13.6
pycparser==2.22
pydantic==2.11.1
pydantic_core==2.33.0
pygame==2.6.1
Pygments==2.19.1
pyjnius==1.6.1
PyJWT==2.10.1
pylint==3.3.4
PySDL2==0.9.17
pytest==8.4.1
pytest-django==4.11.1
python-decouple==3.8
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
schedule==1.2.2
scikit-build==0.18.1
setuptools==79.0.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tomlkit==0.13.2
tqdm==4.67.1
twilio==9.7.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
Werkzeug==3.1.3
wheel==0.45.1
whitenoise==6.9.0
wrapt==1.17.2
yarl==1.20.1
